+++ framework/utils/CHANGELOG.md
@@ -1,3 +1,7 @@
+#3.0.0 (2021-04-23)
+Features
+Session中userId更改为Long类型，并加入traceId
+
 #2.0.0 (2020-06-09)
 Features
 调整包结构