+++ framework/web/CHANGELOG.md
@@ -1,3 +1,7 @@
+#3.0.0 (2021-04-23)
+Features
+Upgrade the 'utils' component version to 2.2.0
+
 #2.2.0 (2021-04-14)
 Features
 Upgrade the 'springBootVersion' to '2.3.1.RELEASE'.